// src/model/superficieChassiParte.ts
export class SuperficieChassiParte {
  id: number;
  superficieChassiID: number;
  forma: number;
  inicioX: number;
  inicioY: number;
  fimX: number;
  fimY: number;
  parteID: number;
  cor: string;
  tipoSelecao: number;
  empresaID: number;
  usuarioID: number;
  bloqueado: boolean;
  dataCadastro: string;
  dataAlteracao: string;
}